admin: Migrate router and navigation stories to @comet/admin package Storybook
Move stories for Router components (ForcePromptRoute, Prompt, SubRoute) and
AppHeader/navigation components (AppHeader, Menu, MenuDynamicVariants,
ThemableAppHeader) from the global storybook into the @comet/admin package
Storybook.

Replace @comet/admin imports with relative imports and remove storyRouterDecorator
(the package Storybook already applies RouterDecorator globally).
