+++ packages/admin/admin/src/appHeader/__stories__/ThemableAppHeader.stories.tsx
@@ -1,4 +1,7 @@
-import { AppHeader, CometLogo, createCometTheme, MuiThemeProvider } from "@comet/admin";
+import { CometLogo } from "../../common/CometLogo";
+import { MuiThemeProvider } from "../../mui/ThemeProvider";
+import { createCometTheme } from "../../theme/createCometTheme";
+import { AppHeader } from "../AppHeader";
 
 export default {
     title: "@comet/admin/theming",
