+++ packages/admin/admin/src/router/__stories__/Prompt.stories.tsx
@@ -1,9 +1,8 @@
-import { ForcePromptRoute, RouterPrompt } from "@comet/admin";
 import { useEffect, useState } from "react";
 import { Redirect, Route, Switch, useLocation } from "react-router";
 import { Link } from "react-router-dom";
 
-import { storyRouterDecorator } from "../../story-router.decorator";
+import { RouterPrompt } from "../Prompt";
 
 function Story() {
     return (
@@ -11,27 +10,18 @@ function Story() {
             <Route path="/foo">
                 <RouterPrompt
                     message={() => {
-                        return "rly?";
+                        return "sure?";
                     }}
-                    subRoutePath="/foo"
+                    subRoutePath="/foo/s"
                 >
-                    <ul>
-                        <li>
-                            <Link to="/foo/sub1">/foo/sub1</Link> (no prompt)
-                        </li>
-                        <li>
-                            <Link to="/foo/sub2">/foo/sub2</Link> (force prompt)
-                        </li>
-                        <li>
-                            <Link to="/foo">/foo</Link> (back)
-                        </li>
-                    </ul>
-                    <Route path="/foo/sub1">
-                        <div>sub1</div>
+                    <Link to="/foo/s/sub">subLink</Link>
+                    <Link to="/foo">fooLink</Link>
+                    <Route path="/foo">
+                        <div>foo</div>
                     </Route>
-                    <ForcePromptRoute path="/foo/sub2">
-                        <div>sub2</div>
-                    </ForcePromptRoute>
+                    <Route path="/foo/s/sub">
+                        <div>sub</div>
+                    </Route>
                 </RouterPrompt>
             </Route>
             <Redirect to="/foo" />
@@ -53,10 +43,9 @@ function Path() {
 
 export default {
     title: "@comet/admin/router",
-    decorators: [storyRouterDecorator()],
 };
 
-export const _ForcePromptRoute = {
+export const NestedRouteWithNonSubPathRouteInPrompt = {
     render: () => {
         return (
             <>
@@ -66,5 +55,5 @@ export const _ForcePromptRoute = {
         );
     },
 
-    name: "ForcePromptRoute",
+    name: "Nested route with non-sub-path route in Prompt",
 };
