+++ packages/admin/admin/src/router/__stories__/SubRoute.stories.tsx
@@ -1,9 +1,8 @@
-import { SubRoute, SubRouteIndexRoute, useSubRoutePrefix } from "@comet/admin";
 import { useEffect, useState } from "react";
 import { Redirect, Route, Switch, useLocation, useRouteMatch } from "react-router";
 import { Link } from "react-router-dom";
 
-import { storyRouterDecorator } from "../../story-router.decorator";
+import { SubRoute, SubRouteIndexRoute, useSubRoutePrefix } from "../SubRoute";
 
 function Cmp1() {
     const urlPrefix = useSubRoutePrefix();
@@ -81,7 +80,6 @@ function Path() {
 
 export default {
     title: "@comet/admin/router",
-    decorators: [storyRouterDecorator()],
 };
 
 export const Subroute = () => {
